FIX: Minor warning about signed / unsigned comparison

+++ Code/Common/itkImageRegionIterator.h
@@ -247,10 +247,10 @@ public:
 
       // Check to see if we are past the last pixel in the region
       // Note that ++ind[0] moves to the next pixel along the row.
-      done = (++ind[0] == startIndex[0] + size[0]);
+      done = (++ind[0] == startIndex[0] + static_cast<long>(size[0]));
       for (unsigned int i=1; done && i < ImageIteratorDimension; i++)
         {
-        done = (ind[i] == startIndex[i] + size[i] - 1);
+        done = (ind[i] == startIndex[i] + static_cast<long>(size[i]) - 1);
         }
       
       // if the iterator is outside the region (but not past region end) then
@@ -259,7 +259,7 @@ public:
       if (!done)
         {
         while ( (dim < ImageIteratorDimension - 1)
-                && (ind[dim] > startIndex[dim] + size[dim] - 1) )
+          && (ind[dim] > startIndex[dim] +  static_cast<long>(size[dim]) - 1) )
           {
           ind[dim] = startIndex[dim];
           ind[++dim]++;
